requirements.txt
streamlit
python-dotenv
openai
pdfplumber
python-pptx
python-docx
reportlab
bcrypt
